ref(docs): address review comments (layout, zod, static ISR)
